+++ crates/cli-tools/CHANGELOG.md
@@ -141,4 +141,4 @@
 
 ## 0.1.0
 
-<!-- Increment to skip CHANGELOG.md test: 13 -->
+<!-- Increment to skip CHANGELOG.md test: 14 -->